+++ packages/conventional-changelog-conventionalcommits/add-bang-notes.js
@@ -1,0 +1,11 @@
+const { breakingHeaderPattern } = require('./parser-opts')()
+
+module.exports = (commit) => {
+  const match = commit.header.match(breakingHeaderPattern)
+  if (match) {
+    const noteText = match[3] // the description of the change.
+    commit.notes.push({
+      text: noteText
+    })
+  }
+}