+++ packages/conventional-changelog-conventionalcommits/parser-opts.js
@@ -1,0 +1,16 @@
+'use strict'
+
+module.exports = function (config) {
+  return {
+    headerPattern: /^(\w*)(?:\((.*)\))?!?: (.*)$/,
+    breakingHeaderPattern: /^(\w*)(?:\((.*)\))?!: (.*)$/,
+    headerCorrespondence: [
+      `type`,
+      `scope`,
+      `subject`
+    ],
+    noteKeywords: [`BREAKING CHANGE`],
+    revertPattern: /^revert:\s([\s\S]*?)\s*This reverts commit (\w*)\./,
+    revertCorrespondence: [`header`, `hash`]
+  }
+}