+++ packages/conventional-recommended-bump/index.js
@@ -29,7 +29,12 @@ function conventionalRecommendedBump (optionsArgument, parserOptsArgument, cbArg
     try {
       presetPackage = conventionalChangelogPresetLoader(options.preset)
     } catch (err) {
-      return cb(new Error(`Unable to load the "${options.preset}" preset package. Please make sure it's installed.`))
+      if (err.message === 'does not exist') {
+        const preset = typeof options.preset === 'object' ? options.preset.name : options.preset
+        return cb(new Error(`Unable to load the "${preset}" preset package. Please make sure it's installed.`))
+      } else {
+        return cb(err)
+      }
     }
   }