+++ packages/conventional-recommended-bump/test/index.spec.js
@@ -21,7 +21,7 @@ betterThanBefore.setups([
   },
   () => { // 2
     fs.writeFileSync(`test1`, ``)
-    shell.exec(`git add --all && git commit -m 'feat: my first commit'`)
+    shell.exec(`git add --all && git commit -m 'feat!: my first commit'`)
   },
   () => { // 3
     shell.exec(`git tag v1.0.0`)
@@ -72,26 +72,28 @@ describe(`conventional-recommended-bump API`, () => {
     })
   })
 
-  describe(`loading a preset package`, () => {
-    it(`throw an error if unable to load a preset package`, done => {
-      preparing(1)
-
-      conventionalRecommendedBump({
-        preset: `does-not-exist`
-      }, {}, err => {
-        assert.ok(err)
-        assert.strictEqual(err.message, `Unable to load the "does-not-exist" preset package. Please make sure it's installed.`)
-        done()
-      })
-    })
-  })
-
   it(`should return an error if there are no commits in the repository`, done => {
     preparing(1)
 
     conventionalRecommendedBump({}, {}, err => {
       assert.ok(err)
       done()
+    })
+  })
+
+  describe('conventionalcommits ! in isolation', () => {
+    it('recommends major if ! is used in isolation', done => {
+      preparing(2)
+
+      conventionalRecommendedBump({
+        preset: {
+          name: 'conventionalcommits'
+        }
+      }, {}, (_, recommendation) => {
+        assert.notStrictEqual(recommendation.reason.indexOf('1 BREAKING'), -1)
+        assert.strictEqual(recommendation.releaseType, 'major')
+        done()
+      })
     })
   })
 
@@ -221,6 +223,49 @@ describe(`conventional-recommended-bump API`, () => {
           done()
         }
       }, () => {})
+    })
+  })
+
+  describe(`loading a preset package`, () => {
+    it(`throws an error if unable to load a preset package`, done => {
+      preparing(5)
+
+      conventionalRecommendedBump({
+        preset: `does-not-exist`
+      }, {}, err => {
+        assert.ok(err)
+        assert.strictEqual(err.message, `Unable to load the "does-not-exist" preset package. Please make sure it's installed.`)
+        done()
+      })
+    })
+
+    it('recommends a minor release when preMajor=true', done => {
+      preparing(5)
+
+      conventionalRecommendedBump({
+        preset: {
+          name: 'conventionalcommits',
+          preMajor: true
+        }
+      }, {}, (_, recommendation) => {
+        assert.notStrictEqual(recommendation.reason.indexOf('1 BREAKING'), -1)
+        assert.strictEqual(recommendation.releaseType, 'minor')
+        done()
+      })
+    })
+
+    it('recommends a major release when preMajor=false', done => {
+      preparing(5)
+
+      conventionalRecommendedBump({
+        preset: {
+          name: 'conventionalcommits'
+        }
+      }, {}, (_, recommendation) => {
+        assert.notStrictEqual(recommendation.reason.indexOf('1 BREAKING'), -1)
+        assert.strictEqual(recommendation.releaseType, 'major')
+        done()
+      })
     })
   })