+++ packages/conventional-changelog-cli/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.20](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.19...conventional-changelog-cli@2.0.20) (2019-05-02)
+
+**Note:** Version bump only for package conventional-changelog-cli
+
+
+
+
+
+## [2.0.19](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.18...conventional-changelog-cli@2.0.19) (2019-05-02)
+
+**Note:** Version bump only for package conventional-changelog-cli
+
+
+
+
+
+## [2.0.18](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.17...conventional-changelog-cli@2.0.18) (2019-04-26)
+
+**Note:** Version bump only for package conventional-changelog-cli
+
+
+
+
+
+## [2.0.17](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.16...conventional-changelog-cli@2.0.17) (2019-04-24)
+
+**Note:** Version bump only for package conventional-changelog-cli
+
+
+
+
+
+## [2.0.16](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.15...conventional-changelog-cli@2.0.16) (2019-04-11)
+
+**Note:** Version bump only for package conventional-changelog-cli
+
+
+
+
+
+## [2.0.15](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.14...conventional-changelog-cli@2.0.15) (2019-04-11)
+
+**Note:** Version bump only for package conventional-changelog-cli
+
+
+
+
+
+## [2.0.14](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.13...conventional-changelog-cli@2.0.14) (2019-04-11)
+
+**Note:** Version bump only for package conventional-changelog-cli
+
+
+
+
+
+## [2.0.13](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.12...conventional-changelog-cli@2.0.13) (2019-04-10)
+
+**Note:** Version bump only for package conventional-changelog-cli
+
+
+
+
+
 ## [2.0.12](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.11...conventional-changelog-cli@2.0.12) (2019-02-14)
 
 **Note:** Version bump only for package conventional-changelog-cli